// frontend/src/api/api8082/apis/FeedbackControllerApi.ts
export interface CreateFeedbackRequest {
    feedbackRequestDTO: FeedbackRequestDTO;
}

export interface DeleteFeedbackRequest {
    id: number;
}

export interface GetFeedbackByIdRequest {
    id: number;
}

/**
 * 
 */
export class FeedbackControllerApi extends runtime.BaseAPI {

    /**
     */
    async createFeedbackRaw(requestParameters: CreateFeedbackRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<FeedbackResponseDTO>> {
        if (requestParameters['feedbackRequestDTO'] == null) {
            throw new runtime.RequiredError(
                'feedbackRequestDTO',
                'Required parameter "feedbackRequestDTO" was null or undefined when calling createFeedback().'
            );
        }

        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};

        headerParameters['Content-Type'] = 'application/json';

        const response = await this.request({
            path: `/feedback`,
            method: 'POST',
            headers: headerParameters,
            query: queryParameters,
            body: FeedbackRequestDTOToJSON(requestParameters['feedbackRequestDTO']),
        }, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => FeedbackResponseDTOFromJSON(jsonValue));
    }

    /**
     */
    async createFeedback(requestParameters: CreateFeedbackRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<FeedbackResponseDTO> {
        const response = await this.createFeedbackRaw(requestParameters, initOverrides);
        return await response.value();
    }

    /**
     */
    async deleteFeedbackRaw(requestParameters: DeleteFeedbackRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<void>> {
        if (requestParameters['id'] == null) {
            throw new runtime.RequiredError(
                'id',
                'Required parameter "id" was null or undefined when calling deleteFeedback().'
            );
        }

        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};

        const response = await this.request({
            path: `/feedback/{id}`.replace(`{${"id"}}`, encodeURIComponent(String(requestParameters['id']))),
            method: 'DELETE',
            headers: headerParameters,
            query: queryParameters,
        }, initOverrides);

        return new runtime.VoidApiResponse(response);
    }

    /**
     */
    async deleteFeedback(requestParameters: DeleteFeedbackRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<void> {
        await this.deleteFeedbackRaw(requestParameters, initOverrides);
    }

    /**
     */
    async getAllFeedbackRaw(initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<Array<FeedbackResponseDTO>>> {
        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};

        const response = await this.request({
            path: `/feedback`,
            method: 'GET',
            headers: headerParameters,
            query: queryParameters,
        }, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => jsonValue.map(FeedbackResponseDTOFromJSON));
    }

    /**
     */
    async getAllFeedback(initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<Array<FeedbackResponseDTO>> {
        const response = await this.getAllFeedbackRaw(initOverrides);
        return await response.value();
    }

    /**
     */
    async getFeedbackByIdRaw(requestParameters: GetFeedbackByIdRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<FeedbackResponseDTO>> {
        if (requestParameters['id'] == null) {
            throw new runtime.RequiredError(
                'id',
                'Required parameter "id" was null or undefined when calling getFeedbackById().'
            );
        }

        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};

        const response = await this.request({
            path: `/feedback/{id}`.replace(`{${"id"}}`, encodeURIComponent(String(requestParameters['id']))),
            method: 'GET',
            headers: headerParameters,
            query: queryParameters,
        }, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => FeedbackResponseDTOFromJSON(jsonValue));
    }

    /**
     */
    async getFeedbackById(requestParameters: GetFeedbackByIdRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<FeedbackResponseDTO> {
        const response = await this.getFeedbackByIdRaw(requestParameters, initOverrides);
        return await response.value();
    }

}