// frontend/src/api/api8082/models/FeedbackRequestDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Authorization server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface FeedbackRequestDTO
 */
export interface FeedbackRequestDTO {
    /**
     * 
     * @type {string}
     * @memberof FeedbackRequestDTO
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof FeedbackRequestDTO
     */
    category?: string;
    /**
     * 
     * @type {string}
     * @memberof FeedbackRequestDTO
     */
    satisfaction?: string;
    /**
     * 
     * @type {boolean}
     * @memberof FeedbackRequestDTO
     */
    subscribe?: boolean;
    /**
     * 
     * @type {string}
     * @memberof FeedbackRequestDTO
     */
    message?: string;
}

/**
 * Check if a given object implements the FeedbackRequestDTO interface.
 */
export function instanceOfFeedbackRequestDTO(value: object): value is FeedbackRequestDTO {
    return true;
}

export function FeedbackRequestDTOFromJSON(json: any): FeedbackRequestDTO {
    return FeedbackRequestDTOFromJSONTyped(json, false);
}

export function FeedbackRequestDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): FeedbackRequestDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'] == null ? undefined : json['title'],
        'category': json['category'] == null ? undefined : json['category'],
        'satisfaction': json['satisfaction'] == null ? undefined : json['satisfaction'],
        'subscribe': json['subscribe'] == null ? undefined : json['subscribe'],
        'message': json['message'] == null ? undefined : json['message'],
    };
}

export function FeedbackRequestDTOToJSON(json: any): FeedbackRequestDTO {
    return FeedbackRequestDTOToJSONTyped(json, false);
}

export function FeedbackRequestDTOToJSONTyped(value?: FeedbackRequestDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'title': value['title'],
        'category': value['category'],
        'satisfaction': value['satisfaction'],
        'subscribe': value['subscribe'],
        'message': value['message'],
    };
}